Refactor display history package and update exception handling.
Renamed "dislay_history" package to "display_history" for consistent naming. Re-introduced DisplayHistoryController with updated logic to incorporate exception handling for recent cities data retrieval. Updated methods and comments in DisplayHistoryPresenter to better reflect functionality, including a new failure view placeholder.
